requirements.txt
psycopg2-binary==2.9.3
pytermgui==6.4.0
PyYAML==6.0
